// src/redux/models/colors.ts
export enum EColors {
  RED = "Red",
  PINK = "Pink",
  PURPLE = "Purple",
  DEEPPURPLE = "DeepPurple",
  INDIGO = "Indigo",
  BLUE = "Blue",
  CYAN = "Cyan",
  TEAL = "Teal",
  GREEN = "Green",
  LIGHTGREEN = "LightGreen",
  LIME = "Lime",
  YELLOW = "Yellow",
  AMBER = "Amber",
  ORANGE = "Orange",
  DEEPORANGE = "DeepOrange",
  BROWN = "Brown"
}